refactor: split confirmation dialog into reusable components, use stable list keys
